Misc Bug Fixes

+++ global_planner/include/global_planner/astar.h
@@ -63,7 +63,7 @@ struct greater1 {
 class AStarExpansion : public Expander {
     public:
         AStarExpansion(PotentialCalculator* p_calc, int nx, int ny);
-        bool calculatePotentials(unsigned char* costs, int start_x, int start_y, int end_x, int end_y, int cycles,
+        bool calculatePotentials(unsigned char* costs, double start_x, double start_y, double end_x, double end_y, int cycles,
                                 float* potential);
     private:
         void add(unsigned char* costs, float* potential, float prev_potential, int next_i, int end_x, int end_y);
